debug

+++ src/gui/domain/ManutARisque.java
@@ -2,7 +2,7 @@ package gui.domain;
 
 public class ManutARisque extends Manutentionnaire implements ARisque {
 
-    public ManutARisque(String firstname, String lastname, int age, String entryYear, int base) {
+    public ManutARisque(String firstname, String lastname, int age, String entryYear, double base) {
         super(firstname, lastname, age, entryYear, base);
     }
